+++ src/api/EntitySetsApi.test.js
@@ -4,7 +4,7 @@ import * as AxiosUtils from '../utils/axios';
 import * as EntitySetsApi from './EntitySetsApi';
 import { ENTITY_SETS_API } from '../constants/ApiNames';
 import { INVALID_PARAMS, INVALID_PARAMS_FOR_OPTIONAL_SS, INVALID_PARAMS_SS } from '../utils/testing/Invalid';
-import { MOCK_ENTITY_SET } from '../utils/testing/MockDataModels';
+import { MOCK_ENTITY_SET } from '../utils/testing/MockData';
 import { genMockBaseUrl, genRandomUUID, getMockAxiosInstance } from '../utils/testing/MockUtils';
 import {
   ALL_PATH,
