+++ src/models/EntitySet.test.js
@@ -1,6 +1,6 @@
 import { Map, Set, fromJS } from 'immutable';
 import EntitySet, { EntitySetBuilder, isValidEntitySet as isValid } from './EntitySet';
-import { MOCK_ENTITY_SET, genRandomEntitySet } from '../utils/testing/MockDataModels';
+import { MOCK_ENTITY_SET, genRandomEntitySet } from '../utils/testing/MockData';
 
 import {
   INVALID_PARAMS,
