+++ src/models/LinkingEntitySet.test.js
@@ -1,6 +1,6 @@
 import LinkingEntitySet, { LinkingEntitySetBuilder, isValidLinkingEntitySet as isValid } from './LinkingEntitySet';
 import { INVALID_PARAMS } from '../utils/testing/Invalid';
-import { MOCK_LINKING_ENTITY_SET_DM } from '../utils/testing/MockDataModels';
+import { MOCK_LINKING_ENTITY_SET_DM } from '../utils/testing/MockData';
 
 describe('LinkingEntitySet', () => {
