+++ src/models/Request.test.js
@@ -1,6 +1,6 @@
 import Request, { RequestBuilder, isValidRequest as isValid } from './Request';
 import { PermissionTypes } from '../constants/types';
-import { MOCK_REQUEST_DM } from '../utils/testing/MockDataModels';
+import { MOCK_REQUEST_DM } from '../utils/testing/MockData';
 
 import {
   INVALID_PARAMS,
