+++ src/models/RequestStatus.test.js
@@ -1,7 +1,7 @@
 import RequestStatus, { RequestStatusBuilder, isValidRequestStatus as isValid } from './RequestStatus';
 import { RequestStateTypes } from '../constants/types';
 import { INVALID_PARAMS, INVALID_PARAMS_SS } from '../utils/testing/Invalid';
-import { MOCK_REQUEST_STATUS_DM } from '../utils/testing/MockDataModels';
+import { MOCK_REQUEST_STATUS_DM } from '../utils/testing/MockData';
 
 describe('RequestStatus', () => {
